requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics and Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# Jupyter (optional, for notebook)
jupyter>=1.0.0
ipykernel>=6.23.0
